+++ Content.Shared/_CP14/MagicSpell/Components/CP14MagicEffectComponent.cs
@@ -1,4 +1,4 @@
-using Content.Shared._CP14.MagicSpell.Interfaces;
+using Content.Shared._CP14.MagicSpell.Spells;
 using Content.Shared.FixedPoint;
 
 namespace Content.Shared._CP14.MagicSpell.Components;
